fix image status storage output signatures
