+++ apps/web/src/app/api/images/status/[id]/route.ts
@@ -2,100 +2,17 @@ import { db } from "@repo/database";
 import { generation } from "@repo/database/schema";
 import { withApiLogging } from "@repo/shared/api-logger";
 import { auth } from "@repo/shared/auth";
-import { generateSignedImageUrl } from "@repo/shared/storage/signed-url";
 import { and, eq } from "drizzle-orm";
 import { type NextRequest, NextResponse } from "next/server";
+import {
+  buildStorageUrl,
+  getImageOutputs,
+  getPromptRepairNotice,
+  getResponseOutput,
+} from "../status-output";
 
 function jsonError(message: string, status = 400) {
   return NextResponse.json({ error: message }, { status });
-}
-
-function asRecord(value: unknown): Record<string, unknown> {
-  return value && typeof value === "object" && !Array.isArray(value)
-    ? (value as Record<string, unknown>)
-    : {};
-}
-
-function stringValue(value: unknown) {
-  return typeof value === "string" && value.trim() ? value.trim() : undefined;
-}
-
-function buildStorageUrl(bucket: string | null, key: string | null) {
-  if (!key) return undefined;
-  const resolvedBucket = bucket || "generations";
-  return generateSignedImageUrl(resolvedBucket, key);
-}
-
-function getImageOutputs(metadata: unknown, bucket: string | null) {
-  const outputImage = asRecord(asRecord(metadata).outputImage);
-  const outputs = outputImage.imageOutputs;
-  const promptRepairNotice = getPromptRepairNotice(metadata);
-  if (!Array.isArray(outputs)) return [];
-  return outputs.flatMap((item, index) => {
-    const output = asRecord(item);
-    const storageKey = stringValue(output.storageKey);
-    const imageUrl =
-      stringValue(output.imageUrl) ||
-      (storageKey ? buildStorageUrl(bucket, storageKey) : undefined);
-    const generationId = stringValue(output.generationId);
-    if (!imageUrl && !generationId) return [];
-    return [
-      {
-        generationId,
-        imageUrl,
-        imageFileId: stringValue(output.imageFileId),
-        webImageMessageId: stringValue(output.webImageMessageId),
-        webImageGroupId: stringValue(output.webImageGroupId),
-        size: stringValue(output.size),
-        revisedPrompt: stringValue(output.revisedPrompt),
-        upstreamRevisedPrompt: stringValue(output.upstreamRevisedPrompt),
-        promptRepairNotice,
-        index,
-        outputRole:
-          output.role === "agent_draft" ||
-          output.role === "choice" ||
-          output.role === "final"
-            ? output.role
-            : undefined,
-      },
-    ];
-  });
-}
-
-function getPromptRepairNotice(metadata: unknown) {
-  const repair = asRecord(asRecord(metadata).moderationPromptRepair);
-  if (repair.succeeded !== true) return undefined;
-  return stringValue(repair.notice) ||
-    "The original prompt was rejected by safety checks, so this request was generated after additional prompt adjustments.";
-}
-
-function getResponseOutput(metadata: unknown) {
-  const output = asRecord(asRecord(metadata).responseOutput);
-  return {
-    responseText: stringValue(output.responseText),
-    responseThinking: stringValue(output.responseThinking),
-    responseAgent: stringValue(output.responseAgent),
-    agentEvents: Array.isArray(output.agentEvents)
-      ? output.agentEvents
-      : undefined,
-    agentRoundCount:
-      typeof output.agentRoundCount === "number"
-        ? output.agentRoundCount
-        : undefined,
-    webConversation:
-      output.webConversation && typeof output.webConversation === "object"
-        ? output.webConversation
-        : undefined,
-    backendMember:
-      output.backendMember && typeof output.backendMember === "object"
-        ? output.backendMember
-        : undefined,
-    responsesPreviousResponse:
-      output.responsesPreviousResponse &&
-      typeof output.responsesPreviousResponse === "object"
-        ? output.responsesPreviousResponse
-        : undefined,
-  };
 }
 
 export const GET = withApiLogging(
